IMMORTAL: Reworks Dialog update loop
Before update() iterated through the dialog message char by char,
returning after each one to logic. Now it tries to do the maximum amonut
of work during one call, like rendering the whole dialog at once if it
is not a scrolling dialog. Certain tokens are handled in Logic::runDialog()
to query user input before advancing to the next dialog or return to the
game for example.